Fixed QuineMcCluskey bugs
